results/rebuttal_new_baselines_llmatic_missing/run-20231111-034200_LLMatic-zero_shot_donnemartin-system-design-oop-eventplanner_0_20-runs_log/eventplanner/11/LLMatic/test_notifications.py
import pytest
from notifications import Notifications

def test_notifications():
	notifications = Notifications()

	# Test adding a notification
	notifications.add_notification('user1', 'event1', 'Notification 1')
	assert notifications.get_notifications('user1') == {'event1': 'Notification 1'}

	# Test getting notifications
	notifications.add_notification('user1', 'event2', 'Notification 2')
	assert notifications.get_notifications('user1') == {'event1': 'Notification 1', 'event2': 'Notification 2'}

	# Test removing a notification
	notifications.remove_notification('user1', 'event1')
	assert notifications.get_notifications('user1') == {'event2': 'Notification 2'}
